Language.Haskell.TH.LanguageExtensions

Description

Language extensions known to GHC

Documentation

data Extension

The language extensions known to GHC.

Note that there is an orphan Binary instance for this type supplied by the GHC.LanguageExtensions module provided by ghc-boot. We can't provide here as this would require adding transitive dependencies to the template-haskell package, which must have a minimal dependency set.

Constructors

Cpp

OverlappingInstances

UndecidableInstances

IncoherentInstances

UndecidableSuperClasses

MonomorphismRestriction

MonoLocalBinds

DeepSubsumption

RelaxedPolyRec

ExtendedDefaultRules

ForeignFunctionInterface

UnliftedFFITypes

InterruptibleFFI

CApiFFI

GHCForeignImportPrim

JavaScriptFFI

ParallelArrays

Arrows

TemplateHaskell

TemplateHaskellQuotes

QualifiedDo

QuasiQuotes

ImplicitParams

ImplicitPrelude

ScopedTypeVariables

AllowAmbiguousTypes

UnboxedTuples

UnboxedSums

UnliftedNewtypes

UnliftedDatatypes

BangPatterns

TypeFamilies

TypeFamilyDependencies

TypeInType

OverloadedStrings

OverloadedLists

NumDecimals

DisambiguateRecordFields

RecordWildCards

NamedFieldPuns

ViewPatterns

GADTs

GADTSyntax

NPlusKPatterns

DoAndIfThenElse

BlockArguments

RebindableSyntax

ConstraintKinds

PolyKinds

DataKinds

TypeData

InstanceSigs

ApplicativeDo

LinearTypes

StandaloneDeriving

DeriveDataTypeable

AutoDeriveTypeable

DeriveFunctor

DeriveTraversable

DeriveFoldable

DeriveGeneric

DefaultSignatures

DeriveAnyClass

DeriveLift

DerivingStrategies

DerivingVia

TypeSynonymInstances

FlexibleContexts

FlexibleInstances

ConstrainedClassMethods

MultiParamTypeClasses

NullaryTypeClasses

FunctionalDependencies

UnicodeSyntax

ExistentialQuantification

MagicHash

EmptyDataDecls

KindSignatures

RoleAnnotations

ParallelListComp

TransformListComp

MonadComprehensions

GeneralizedNewtypeDeriving

RecursiveDo

PostfixOperators

TupleSections

PatternGuards

LiberalTypeSynonyms

RankNTypes

ImpredicativeTypes

TypeOperators

ExplicitNamespaces

PackageImports

ExplicitForAll

AlternativeLayoutRule

AlternativeLayoutRuleTransitional

DatatypeContexts

NondecreasingIndentation

RelaxedLayout

TraditionalRecordSyntax

LambdaCase

MultiWayIf

BinaryLiterals

NegativeLiterals

HexFloatLiterals

DuplicateRecordFields

OverloadedLabels

EmptyCase

PatternSynonyms

PartialTypeSignatures

NamedWildCards

StaticPointers

TypeApplications

Strict

StrictData

EmptyDataDeriving

NumericUnderscores

QuantifiedConstraints

StarIsType

ImportQualifiedPost

CUSKs

StandaloneKindSignatures

LexicalNegation

FieldSelectors

OverloadedRecordDot

OverloadedRecordUpdate