Language.Haskell.TH.Syntax

Description

Abstract syntax definitions for Template Haskell.

Documentation

data Foreign

Constructors

ImportF Callconv Safety String Name Type

ExportF Callconv String Name Type

data Type

Constructors

ForallT [TyVarBndr Specificity] Cxt Type
forall <vars>. <ctxt> => <type>

ForallVisT [TyVarBndr ()] Type
forall <vars> -> <type>

AppT Type Type
T a b

AppKindT Type Kind
T @k t

SigT Type Kind
t :: k

VarT Name
a

ConT Name
T

PromotedT Name
'T

InfixT Type Name Type
T + T

UInfixT Type Name Type
T + T

See Language.Haskell.TH.Syntax

PromotedInfixT Type Name Type
T :+: T

PromotedUInfixT Type Name Type
T :+: T

See Language.Haskell.TH.Syntax

ParensT Type
(T)

TupleT Int
(,), (,,), etc.

UnboxedTupleT Int
(#,#), (#,,#), etc.

UnboxedSumT SumArity
(#|#), (#||#), etc.

ArrowT
->

MulArrowT
%n ->

Generalised arrow type with multiplicity argument

EqualityT
~

ListT
[]

PromotedTupleT Int
'(), '(,), '(,,), etc.

PromotedNilT
'[]

PromotedConsT
'(:)

StarT
*

ConstraintT
Constraint

LitT TyLit
0, 1, 2, etc.

WildCardT
_

ImplicitParamT String Type
?x :: t

data Module

Obtained from reifyModule and thisModule.

Constructors

Module PkgName ModName

type Unlifted = Bool

In PrimTyConI, is the type constructor unlifted?

data DecidedStrictness

Unlike SourceStrictness and SourceUnpackedness, DecidedStrictness refers to the strictness annotations that the compiler chooses for a data constructor field, which may be different from what is written in source code.

Note that non-unpacked strict fields are assigned DecidedLazy when a bang would be inappropriate, such as the field of a newtype constructor and fields that have an unlifted type.

See reifyConStrictness for more information.

Constructors

DecidedLazy
Field inferred to not have a bang.

DecidedStrict
Field inferred to have a bang.

DecidedUnpack
Field inferred to be unpacked.

data SourceStrictness

SourceStrictness corresponds to strictness annotations found in the source code.

This may not agree with the annotations returned by reifyConStrictness. See reifyConStrictness for more information.

Constructors

NoSourceStrictness
C a

SourceLazy
C {~}a

SourceStrict
C {!}a

data SourceUnpackedness

SourceUnpackedness corresponds to unpack annotations found in the source code.

This may not agree with the annotations returned by reifyConStrictness. See reifyConStrictness for more information.

Constructors

NoSourceUnpackedness
C a

SourceNoUnpack
C { {-# NOUNPACK #-} } a

SourceUnpack
C { {-# UNPACK #-} } a

data Fixity

Constructors

Fixity Int FixityDirection

data Con

A single data constructor.

The constructors for Con can roughly be divided up into two categories: those for constructors with "vanilla" syntax (NormalC, RecC, and InfixC), and those for constructors with GADT syntax (GadtC and RecGadtC). The ForallC constructor, which quantifies additional type variables and class contexts, can surround either variety of constructor. However, the type variables that it quantifies are different depending on what constructor syntax is used:

If a ForallC surrounds a constructor with vanilla syntax, then the ForallC will only quantify existential type variables. For example:

  data Foo a = forall b. MkFoo a b
  

In MkFoo, ForallC will quantify b, but not a.

If a ForallC surrounds a constructor with GADT syntax, then the ForallC will quantify all type variables used in the constructor. For example:

  data Bar a b where
    MkBar :: (a ~ b) => c -> MkBar a b
  

In MkBar, ForallC will quantify a, b, and c.

Multiplicity annotations for data types are currently not supported in Template Haskell (i.e. all fields represented by Template Haskell will be linear).

Constructors

NormalC Name [BangType]
C Int a

RecC Name [VarBangType]
C { v :: Int, w :: a }

InfixC BangType Name BangType
Int :+ a

ForallC [TyVarBndr Specificity] Cxt Con
forall a. Eq a => C [a]

GadtC [Name] [BangType] Type
C :: a -> b -> T b Int

RecGadtC [Name] [VarBangType] Type
C :: { v :: Int } -> T b Int

type Strict = Bang

As of template-haskell-2.11.0.0, Strict has been replaced by Bang.

newtype Q a

Constructors

Q

Fields

unQ :: forall m. Quasi m => m a

class Monad m => Quote m where

The Quote class implements the minimal interface which is necessary for desugaring quotations.

The Monad m superclass is needed to stitch together the different AST fragments.

newName is used when desugaring binding structures such as lambdas to generate fresh names.

Therefore the type of an untyped quotation in GHC is `Quote m => m Exp`

For many years the type of a quotation was fixed to be `Q Exp` but by more precisely specifying the minimal interface it enables the Exp to be extracted purely from the quotation without interacting with Q.

Methods

newName :: String -> m Name

Generate a fresh name, which cannot be captured.

For example, this:

f = $(do
    nm1 <- newName "x"
    let nm2 = mkName "x"
    return (LamE [VarP nm1] (LamE [VarP nm2] (VarE nm1)))
   )

will produce the splice

f = \x0 -> \x -> x0

In particular, the occurrence VarE nm1 refers to the binding VarP nm1, and is not captured by the binding VarP nm2.

Although names generated by newName cannot be captured, they can capture other names. For example, this:

g = $(do
  nm1 <- newName "x"
  let nm2 = mkName "x"
  return (LamE [VarP nm2] (LamE [VarP nm1] (VarE nm2)))
 )

will produce the splice

g = \x -> \x0 -> x0

since the occurrence VarE nm2 is captured by the innermost binding of x, namely VarP nm1.

data Loc

Constructors

Loc

Fields

loc_filename :: String

loc_package :: String

loc_module :: String

loc_start :: CharPos

loc_end :: CharPos

data Info

Obtained from reify in the Q Monad.

Constructors

ClassI Dec [InstanceDec]
A class, with a list of its visible instances

ClassOpI Name Type ParentName
A class method

TyConI Dec
A "plain" type constructor. "Fancier" type constructors are returned using PrimTyConI or FamilyI as appropriate. At present, this reified declaration will never have derived instances attached to it (if you wish to check for an instance, see reifyInstances).

FamilyI Dec [InstanceDec]
A type or data family, with a list of its visible instances. A closed type family is returned with 0 instances.

PrimTyConI Name Arity Unlifted
A "primitive" type constructor, which can't be expressed with a Dec. Examples: (->), Int#.

DataConI Name Type ParentName
A data constructor

PatSynI Name PatSynType
A pattern synonym

VarI Name Type (Maybe Dec)
A "value" variable (as opposed to a type variable, see TyVarI).

The Maybe Dec field contains Just the declaration which defined the variable - including the RHS of the declaration - or else Nothing, in the case where the RHS is unavailable to the compiler. At present, this value is always Nothing: returning the RHS has not yet been implemented because of lack of interest.

TyVarI Name Type
A type variable.

The Type field contains the type which underlies the variable. At present, this is always VarT theName, but future changes may permit refinement of this.

data ModuleInfo

Obtained from reifyModule in the Q Monad.

Constructors

ModuleInfo [Module]
Contains the import list of the module.

type InstanceDec = Dec

InstanceDec describes a single instance of a class or type function. It is just a Dec, but guaranteed to be one of the following:

InstanceD (with empty [Dec])

DataInstD or NewtypeInstD (with empty derived [Name])

TySynInstD

type ParentName = Name

In ClassOpI and DataConI, name of the parent class or type

type SumAlt = Int

In UnboxedSumE and UnboxedSumP, the number associated with a particular data constructor. SumAlts are one-indexed and should never exceed the value of its corresponding SumArity. For example:

(#_|#) has SumAlt 1 (out of a total SumArity of 2)

(#|_#) has SumAlt 2 (out of a total SumArity of 2)

type SumArity = Int

In UnboxedSumE, UnboxedSumT, and UnboxedSumP, the total number of SumAlts. For example, (#|#) has a SumArity of 2.

type Arity = Int

In PrimTyConI, arity of the type constructor

data AnnLookup

Annotation target for reifyAnnotations

Constructors

AnnLookupModule Module

AnnLookupName Name

newtype TExp (a :: TYPE (r :: RuntimeRep))

Represents an expression which has type a. Built on top of Exp, typed expressions allow for type-safe splicing via:

typed quotes, written as [|| ... ||] where ... is an expression; if that expression has type a, then the quotation has type Q (TExp a)

typed splices inside of typed quotes, written as $$(...) where ... is an arbitrary expression of type Q (TExp a)

Traditional expression quotes and splices let us construct ill-typed expressions:

>>> fmap ppr $ runQ [| True == $( [| "foo" |] ) |]
GHC.Types.True GHC.Classes.== "foo"
>>> GHC.Types.True GHC.Classes.== "foo"
<interactive> error:
    • Couldn't match expected type ‘Bool’ with actual type ‘[Char]’
    • In the second argument of ‘(==)’, namely ‘"foo"’
      In the expression: True == "foo"
      In an equation for ‘it’: it = True == "foo"

With typed expressions, the type error occurs when constructing the Template Haskell expression:

>>> fmap ppr $ runQ [|| True == $$( [|| "foo" ||] ) ||]
<interactive> error:
    • Couldn't match type ‘[Char]’ with ‘Bool’
      Expected type: Q (TExp Bool)
        Actual type: Q (TExp [Char])
    • In the Template Haskell quotation [|| "foo" ||]
      In the expression: [|| "foo" ||]
      In the Template Haskell splice $$([|| "foo" ||])

Representation-polymorphic since template-haskell-2.16.0.0.

Constructors

TExp

Fields

unType :: Exp

Underlying untyped Template Haskell expression

newtype Code m (a :: TYPE (r :: RuntimeRep))

Constructors

Code

Fields

examineCode :: m (TExp a)

Underlying monadic value

data Name

An abstract type representing names in the syntax tree.

Names can be constructed in several ways, which come with different name-capture guarantees (see Language.Haskell.TH.Syntax for an explanation of name capture):

the built-in syntax 'f and ''T can be used to construct names, The expression 'f gives a Name which refers to the value f currently in scope, and ''T gives a Name which refers to the type T currently in scope. These names can never be captured.

lookupValueName and lookupTypeName are similar to 'f and ''T respectively, but the Names are looked up at the point where the current splice is being run. These names can never be captured.

newName monadically generates a new name, which can never be captured.

mkName generates a capturable name.

Names constructed using newName and mkName may be used in bindings (such as let x = ... or x -> ...), but names constructed using lookupValueName, lookupTypeName, 'f, ''T may not.

Constructors

Name OccName NameFlavour

data NameSpace

Constructors

VarName
Variables

DataName
Data constructors

TcClsName
Type constructors and classes; Haskell has them in the same name space for now.

data Dec

Constructors

FunD Name [Clause]
{ f p1 p2 = b where decs }

ValD Pat Body [Dec]
{ p = b where decs }

DataD Cxt Name [TyVarBndr ()] (Maybe Kind) [Con] [DerivClause]
{ data Cxt x => T x = A x | B (T x)
       deriving (Z,W)
       deriving stock Eq }

NewtypeD Cxt Name [TyVarBndr ()] (Maybe Kind) Con [DerivClause]
{ newtype Cxt x => T x = A (B x)
       deriving (Z,W Q)
       deriving stock Eq }

TypeDataD Name [TyVarBndr ()] (Maybe Kind) [Con]
{ type data T x = A x | B (T x) }

TySynD Name [TyVarBndr ()] Type
{ type T x = (x,x) }

ClassD Cxt Name [TyVarBndr ()] [FunDep] [Dec]
{ class Eq a => Ord a where ds }

InstanceD (Maybe Overlap) Cxt Type [Dec]
{ instance {-# OVERLAPS #-}
        Show w => Show [w] where ds }

SigD Name Type
{ length :: [a] -> Int }

KiSigD Name Kind
{ type TypeRep :: k -> Type }

ForeignD Foreign
{ foreign import ... }
{ foreign export ... }

InfixD Fixity Name
{ infix 3 foo }

DefaultD [Type]
{ default (Integer, Double) }

PragmaD
pragmas

Fields

Pragma

{ {-# INLINE [1] foo #-} }

DataFamilyD
data families (may also appear in [Dec] of ClassD and InstanceD)

Fields

Name

[TyVarBndr ()]

(Maybe Kind)

{ data family T a b c :: * }

DataInstD Cxt (Maybe [TyVarBndr ()]) Type (Maybe Kind) [Con] [DerivClause]
{ data instance Cxt x => T [x]
       = A x | B (T x)
       deriving (Z,W)
       deriving stock Eq }

NewtypeInstD Cxt (Maybe [TyVarBndr ()]) Type (Maybe Kind) Con [DerivClause]
{ newtype instance Cxt x => T [x]
        = A (B x)
        deriving (Z,W)
        deriving stock Eq }

TySynInstD TySynEqn
{ type instance ... }

OpenTypeFamilyD
open type families (may also appear in [Dec] of ClassD and InstanceD)

Fields

TypeFamilyHead

{ type family T a b c = (r :: *) | r -> a b }

ClosedTypeFamilyD TypeFamilyHead [TySynEqn]
{ type family F a b = (r :: *) | r -> a where ... }

RoleAnnotD Name [Role]
{ type role T nominal representational }

StandaloneDerivD (Maybe DerivStrategy) Cxt Type
{ deriving stock instance Ord a => Ord (Foo a) }

DefaultSigD Name Type
{ default size :: Data a => a -> Int }

PatSynD
Pattern Synonyms

Fields

Name

PatSynArgs

PatSynDir

Pat

{ pattern P v1 v2 .. vn <- p } unidirectional or { pattern P v1 v2 .. vn = p } implicit bidirectional or { pattern P v1 v2 .. vn <- p where P v1 v2 .. vn = e } explicit bidirectional

also, besides prefix pattern synonyms, both infix and record pattern synonyms are supported. See PatSynArgs for details

PatSynSigD Name PatSynType
A pattern synonym's type signature.

ImplicitParamBindD String Exp
{ ?x = expr }

Implicit parameter binding declaration. Can only be used in let and where clauses which consist entirely of implicit bindings.

data Clause

Constructors

Clause [Pat] Body [Dec]
f { p1 p2 = body where decs }

data Bang

Constructors

Bang SourceUnpackedness SourceStrictness
C { {-# UNPACK #-} !}a

data Callconv

Constructors

CCall

StdCall

CApi

Prim

JavaScript

data Safety

Constructors

Unsafe

Safe

Interruptible

data Pragma

Constructors

InlineP Name Inline RuleMatch Phases

OpaqueP Name

SpecialiseP Name Type (Maybe Inline) Phases

SpecialiseInstP Type

RuleP String (Maybe [TyVarBndr ()]) [RuleBndr] Exp Exp Phases

AnnP AnnTarget Exp

LineP Int String

CompleteP [Name] (Maybe Name)
{ {-# COMPLETE C_1, ..., C_i [ :: T ] #-} }

data Inline

Constructors

NoInline

Inline

Inlinable

data RuleMatch

Constructors

ConLike

FunLike

data Phases

Constructors

AllPhases

FromPhase Int

BeforePhase Int

data RuleBndr

Constructors

RuleVar Name

TypedRuleVar Name Type

data AnnTarget

Constructors

ModuleAnnotation

TypeAnnotation Name

ValueAnnotation Name

data FunDep

Constructors

FunDep [Name] [Name]

data TySynEqn

One equation of a type family instance or closed type family. The arguments are the left-hand-side type and the right-hand-side result.

For instance, if you had the following type family:

type family Foo (a :: k) :: k where
  forall k (a :: k). Foo @k a = a

The Foo @k a = a equation would be represented as follows:

TySynEqn (Just [PlainTV k, KindedTV a (VarT k)])
           (AppT (AppKindT (ConT ''Foo) (VarT k)) (VarT a))
           (VarT a)

Constructors

TySynEqn (Maybe [TyVarBndr ()]) Type Type

data TypeFamilyHead

Common elements of OpenTypeFamilyD and ClosedTypeFamilyD. By analogy with "head" for type classes and type class instances as defined in Type classes: an exploration of the design space, the TypeFamilyHead is defined to be the elements of the declaration between type family and where.

Constructors

TypeFamilyHead Name [TyVarBndr ()] FamilyResultSig (Maybe InjectivityAnn)

data FixityDirection

Constructors

InfixL

InfixR

InfixN

data PatSynDir

A pattern synonym's directionality.

Constructors

Unidir
pattern P x {<-} p

ImplBidir
pattern P x {=} p

ExplBidir [Clause]
pattern P x {<-} p where P x = e

data PatSynArgs

A pattern synonym's argument type.

Constructors

PrefixPatSyn [Name]
pattern P {x y z} = p

InfixPatSyn Name Name
pattern {x P y} = p

RecordPatSyn [Name]
pattern P { {x,y,z} } = p

data Exp

Constructors

VarE Name
{ x }

ConE Name
data T1 = C1 t1 t2; p = {C1} e1 e2

LitE Lit
{ 5 or 'c'}

AppE Exp Exp
{ f x }

AppTypeE Exp Type
{ f @Int }

InfixE (Maybe Exp) Exp (Maybe Exp)
{x + y} or {(x+)} or {(+ x)} or {(+)}

UInfixE Exp Exp Exp
{x + y}

See Language.Haskell.TH.Syntax

ParensE Exp
{ (e) }

See Language.Haskell.TH.Syntax

LamE [Pat] Exp
{ \ p1 p2 -> e }

LamCaseE [Match]
{ \case m1; m2 }

LamCasesE [Clause]
{ \cases m1; m2 }

TupE [Maybe Exp]
{ (e1,e2) }

The Maybe is necessary for handling tuple sections.

(1,)

translates to

TupE [Just (LitE (IntegerL 1)),Nothing]

UnboxedTupE [Maybe Exp]
{ (# e1,e2 #) }

The Maybe is necessary for handling tuple sections.

(# 'c', #)

translates to

UnboxedTupE [Just (LitE (CharL 'c')),Nothing]

UnboxedSumE Exp SumAlt SumArity
{ (#|e|#) }

CondE Exp Exp Exp
{ if e1 then e2 else e3 }

MultiIfE [(Guard, Exp)]
{ if | g1 -> e1 | g2 -> e2 }

LetE [Dec] Exp
{ let { x=e1; y=e2 } in e3 }

CaseE Exp [Match]
{ case e of m1; m2 }

DoE (Maybe ModName) [Stmt]
{ do { p <- e1; e2 } } or a qualified do if the module name is present

MDoE (Maybe ModName) [Stmt]
{ mdo { x <- e1 y; y <- e2 x; } } or a qualified mdo if the module name is present

CompE [Stmt]
{ [ (x,y) | x <- xs, y <- ys ] }

The result expression of the comprehension is the last of the Stmts, and should be a NoBindS.

E.g. translation:

[ f x | x <- xs ]

CompE [BindS (VarP x) (VarE xs), NoBindS (AppE (VarE f) (VarE x))]

ArithSeqE Range
{ [ 1 ,2 .. 10 ] }

ListE [Exp]
{ [1,2,3] }

SigE Exp Type
{ e :: t }

RecConE Name [FieldExp]
{ T { x = y, z = w } }

RecUpdE Exp [FieldExp]
{ (f x) { z = w } }

StaticE Exp
{ static e }

UnboundVarE Name
{ _x }

This is used for holes or unresolved identifiers in AST quotes. Note that it could either have a variable name or constructor name.

LabelE String
{ #x } ( Overloaded label )

ImplicitParamVarE String
{ ?x } ( Implicit parameter )

GetFieldE Exp String
{ exp.field } ( Overloaded Record Dot )

ProjectionE (NonEmpty String)
(.x) or (.x.y) (Record projections)

data Match

Constructors

Match Pat Body [Dec]
case e of { pat -> body where decs }

data Body

Constructors

GuardedB [(Guard, Exp)]
f p { | e1 = e2
      | e3 = e4 }
 where ds

NormalB Exp
f p { = e } where ds

data Guard

Constructors

NormalG Exp
f x { | odd x } = x

PatG [Stmt]
f x { | Just y <- x, Just z <- y } = z

data Stmt

Constructors

BindS Pat Exp
p <- e

LetS [Dec]
{ let { x=e1; y=e2 } }

NoBindS Exp
e

ParS [[Stmt]]
x <- e1 | s2, s3 | s4 (in CompE)

RecS [Stmt]
rec { s1; s2 }

data Range

Constructors

FromR Exp

FromThenR Exp Exp

FromToR Exp Exp

FromThenToR Exp Exp Exp

data Lit

Constructors

CharL Char

StringL String

IntegerL Integer
Used for overloaded and non-overloaded literals. We don't have a good way to represent non-overloaded literals at the moment. Maybe that doesn't matter?

RationalL Rational

IntPrimL Integer

WordPrimL Integer

FloatPrimL Rational

DoublePrimL Rational

StringPrimL [Word8]
A primitive C-style string, type Addr#

BytesPrimL Bytes
Some raw bytes, type Addr#:

CharPrimL Char

data Pat

Pattern in Haskell given in {}

Constructors

LitP Lit
{ 5 or 'c' }

VarP Name
{ x }

TupP [Pat]
{ (p1,p2) }

UnboxedTupP [Pat]
{ (# p1,p2 #) }

UnboxedSumP Pat SumAlt SumArity
{ (#|p|#) }

ConP Name [Type] [Pat]
data T1 = C1 t1 t2; {C1 @ty1 p1 p2} = e

InfixP Pat Name Pat
foo ({x :+ y}) = e

UInfixP Pat Name Pat
foo ({x :+ y}) = e

See Language.Haskell.TH.Syntax

ParensP Pat
{(p)}

See Language.Haskell.TH.Syntax

TildeP Pat
{ ~p }

BangP Pat
{ !p }

AsP Name Pat
{ x @ p }

WildP
{ _ }

RecP Name [FieldPat]
f (Pt { pointx = x }) = g x

ListP [Pat]
{ [1,2,3] }

SigP Pat Type
{ p :: t }

ViewP Exp Pat
{ e -> p }

type FieldExp = (Name, Exp)

type FieldPat = (Name, Pat)

data TyVarBndr flag

Constructors

PlainTV Name flag
a

KindedTV Name flag Kind
(a :: k)

data TyLit

Constructors

NumTyLit Integer
2

StrTyLit String
"Hello"

CharTyLit Char
'C', @since 4.16.0.0

type Kind = Type

To avoid duplication between kinds and types, they are defined to be the same. Naturally, you would never have a type be StarT and you would never have a kind be SigT, but many of the other constructors are shared. Note that the kind Bool is denoted with ConT, not PromotedT. Similarly, tuple kinds are made with TupleT, not PromotedTupleT.

type Cxt

Arguments

= [Pred]
(Eq a, Ord b)

type Pred = Type

Since the advent of ConstraintKinds, constraints are really just types. Equality constraints use the EqualityT constructor. Constraints may also be tuples of other constraints.

data Role

Role annotations

Constructors

NominalR
nominal

RepresentationalR
representational

PhantomR
phantom

InferR
_

data Specificity

Constructors

SpecifiedSpec
a

InferredSpec
{a}

data FamilyResultSig

Type family result signature

Constructors

NoSig
no signature

KindSig Kind
k

TyVarSig (TyVarBndr ())
= r, = (r :: k)

data InjectivityAnn

Injectivity annotation

Constructors

InjectivityAnn Name [Name]

type PatSynType = Type

A pattern synonym's type. Note that a pattern synonym's fully specified type has a peculiar shape coming with two forall quantifiers and two constraint contexts. For example, consider the pattern synonym

pattern P x1 x2 ... xn = <some-pattern>

P's complete type is of the following form

pattern P :: forall universals.   required constraints
          => forall existentials. provided constraints
          => t1 -> t2 -> ... -> tn -> t

consisting of four parts:

the (possibly empty lists of) universally quantified type variables and required constraints on them.

the (possibly empty lists of) existentially quantified type variables and the provided constraints on them.

the types t1, t2, .., tn of x1, x2, .., xn, respectively

the type t of <some-pattern>, mentioning only universals.

Pattern synonym types interact with TH when (a) reifying a pattern synonym, (b) pretty printing, or (c) specifying a pattern synonym's type signature explicitly:

Reification always returns a pattern synonym's fully specified type in abstract syntax.

Pretty printing via pprPatSynType abbreviates a pattern synonym's type unambiguously in concrete syntax: The rule of thumb is to print initial empty universals and the required context as () =>, if existentials and a provided context follow. If only universals and their required context, but no existentials are specified, only the universals and their required context are printed. If both or none are specified, so both (or none) are printed.

When specifying a pattern synonym's type explicitly with PatSynSigD either one of the universals, the existentials, or their contexts may be left empty.

See the GHC user's guide for more information on pattern synonyms and their types: https://downloads.haskell.org/~ghc/latest/docs/html/users_guide/glasgow_exts.html#pattern-synonyms.

type BangType = (Bang, Type)

type VarBangType = (Name, Bang, Type)

data DocLoc

A location at which to attach Haddock documentation. Note that adding documentation to a Name defined oustide of the current module will cause an error.

Constructors

ModuleDoc
At the current module's header.

DeclDoc Name
At a declaration, not necessarily top level.

ArgDoc Name Int
At a specific argument of a function, indexed by its position.

InstDoc Type
At a class or family instance.

data DerivClause

A single deriving clause at the end of a datatype.

Constructors

DerivClause (Maybe DerivStrategy) Cxt
{ deriving stock (Eq, Ord) }

data DerivStrategy

What the user explicitly requests when deriving an instance.

Constructors

StockStrategy
A "standard" derived instance

AnyclassStrategy
-XDeriveAnyClass

NewtypeStrategy
-XGeneralizedNewtypeDeriving

ViaStrategy Type
-XDerivingVia

data Overlap

Varieties of allowed instance overlap.

Constructors

Overlappable
May be overlapped by more specific instances

Overlapping
May overlap a more general instance

Overlaps
Both Overlapping and Overlappable

Incoherent
Both Overlapping and Overlappable, and pick an arbitrary one if multiple choices are available.

type Uniq = Integer

Uniq is used by GHC to distinguish names from each other.

data NameFlavour

Constructors

NameS
An unqualified name; dynamically bound

NameQ ModName
A qualified name; dynamically bound

NameU !Uniq
A unique local name

NameL !Uniq
Local name bound outside of the TH AST

NameG NameSpace PkgName ModName
Global name bound outside of the TH AST: An original name (occurrences only, not binders) Need the namespace too to be sure which thing we are naming

data NameIs

Constructors

Alone

Applied

Infix

data Bytes

Raw bytes embedded into the binary.

Avoid using Bytes constructor directly as it is likely to change in the future. Use helpers such as mkBytes in Language.Haskell.TH.Lib instead.

Constructors

Bytes

Fields

bytesPtr :: ForeignPtr Word8

Pointer to the data

bytesOffset :: Word

Offset from the pointer

bytesSize :: Word

Number of bytes

class (MonadIO m, MonadFail m) => Quasi m where

Minimal complete definition

qNewName, qReport, qRecover, qLookupName, qReify, qReifyFixity, qReifyType, qReifyInstances, qReifyRoles, qReifyAnnotations, qReifyModule, qReifyConStrictness, qLocation, qGetPackageRoot, qAddDependentFile, qAddTempFile, qAddTopDecls, qAddForeignFilePath, qAddModFinalizer, qAddCorePlugin, qGetQ, qPutQ, qIsExtEnabled, qExtsEnabled, qPutDoc, qGetDoc

Methods

qNewName

Arguments

:: String

-> m Name
Fresh names

qReport

Arguments

:: Bool

-> String

-> m ()
Report an error (True) or warning (False) ...but carry on; use fail to stop

qRecover

Arguments

:: m a
the error handler

-> m a
action which may fail

-> m a
Recover from the monadic fail

qLookupName :: Bool -> String -> m (Maybe Name)

qReify :: Name -> m Info

qReifyFixity :: Name -> m (Maybe Fixity)

qReifyType :: Name -> m Type

qReifyInstances :: Name -> [Type] -> m [Dec]

qReifyRoles :: Name -> m [Role]

qReifyAnnotations :: Data a => AnnLookup -> m [a]

qReifyModule :: Module -> m ModuleInfo

qReifyConStrictness :: Name -> m [DecidedStrictness]

qLocation :: m Loc

qRunIO :: IO a -> m a

qGetPackageRoot :: m FilePath

qAddDependentFile :: FilePath -> m ()

qAddTempFile :: String -> m FilePath

qAddTopDecls :: [Dec] -> m ()

qAddForeignFilePath :: ForeignSrcLang -> String -> m ()

qAddModFinalizer :: Q () -> m ()

qAddCorePlugin :: String -> m ()

qGetQ :: Typeable a => m (Maybe a)

qPutQ :: Typeable a => a -> m ()

qIsExtEnabled :: Extension -> m Bool

qExtsEnabled :: m [Extension]

qPutDoc :: DocLoc -> String -> m ()

qGetDoc :: DocLoc -> m (Maybe String)

class Lift (t :: TYPE r) where

A Lift instance can have any of its values turned into a Template Haskell expression. This is needed when a value used within a Template Haskell quotation is bound outside the Oxford brackets ([| ... |] or [|| ... ||]) but not at the top level. As an example:

add1 :: Int -> Q (TExp Int)
add1 x = [|| x + 1 ||]

Template Haskell has no way of knowing what value x will take on at splice-time, so it requires the type of x to be an instance of Lift.

A Lift instance must satisfy $(lift x) ≡ x and $$(liftTyped x) ≡ x for all x, where $(...) and $$(...) are Template Haskell splices. It is additionally expected that lift x ≡ unTypeQ (liftTyped x).

Lift instances can be derived automatically by use of the -XDeriveLift GHC language extension:

{-# LANGUAGE DeriveLift #-}
module Foo where

import Language.Haskell.TH.Syntax

data Bar a = Bar1 a (Bar a) | Bar2 String
  deriving Lift

Representation-polymorphic since template-haskell-2.16.0.0.

Minimal complete definition

liftTyped

Methods

lift :: Quote m => t -> m Exp

Turn a value into a Template Haskell expression, suitable for use in a splice.

default lift :: (r ~ 'BoxedRep 'Lifted, Quote m) => t -> m Exp

liftTyped :: Quote m => t -> Code m t

Turn a value into a Template Haskell typed expression, suitable for use in a typed splice.

Since: template-haskell-2.16.0.0

newtype ModName

Constructors

ModName String

newtype PkgName

Constructors

PkgName String

newtype OccName

Constructors

OccName String

type CharPos

Arguments

= (Int, Int)
Line and character position

type StrictType = BangType

As of template-haskell-2.11.0.0, StrictType has been replaced by BangType.

type VarStrictType = VarBangType

As of template-haskell-2.11.0.0, VarStrictType has been replaced by VarBangType.

recover

Arguments

:: Q a
handler to invoke on failure

-> Q a
computation to run

-> Q a

Recover from errors raised by reportError or fail.

reportError :: String -> Q ()

Report an error to the user, but allow the current splice's computation to carry on. To abort the computation, use fail.

runIO :: IO a -> Q a

The runIO function lets you run an I/O computation in the Q monad. Take care: you are guaranteed the ordering of calls to runIO within a single Q computation, but not about the order in which splices are run.

Note: for various murky reasons, stdout and stderr handles are not necessarily flushed when the compiler finishes running, so you should flush them yourself.

runQ :: Quasi m => Q a -> m a

reportWarning :: String -> Q ()

Report a warning to the user, and carry on.

report :: Bool -> String -> Q ()

Deprecated: Use reportError or reportWarning instead

Report an error (True) or warning (False), but carry on; use fail to stop.

location :: Q Loc

The location at which this computation is spliced.

reify :: Name -> Q Info

reify looks up information about the Name. It will fail with a compile error if the Name is not visible. A Name is visible if it is imported or defined in a prior top-level declaration group. See the documentation for newDeclarationGroup for more details.

It is sometimes useful to construct the argument name using lookupTypeName or lookupValueName to ensure that we are reifying from the right namespace. For instance, in this context:

data D = D

which D does reify (mkName "D") return information about? (Answer: D-the-type, but don't rely on it.) To ensure we get information about D-the-value, use lookupValueName:

do
  Just nm <- lookupValueName "D"
  reify nm

and to get information about D-the-type, use lookupTypeName.

reifyModule :: Module -> Q ModuleInfo

reifyModule mod looks up information about module mod. To look up the current module, call this function with the return value of thisModule.

newDeclarationGroup :: Q [Dec]

Template Haskell is capable of reifying information about types and terms defined in previous declaration groups. Top-level declaration splices break up declaration groups.

For an example, consider this code block. We define a datatype X and then try to call reify on the datatype.

module Check where

data X = X
    deriving Eq

$(do
    info <- reify ''X
    runIO $ print info
 )

This code fails to compile, noting that X is not available for reification at the site of reify. We can fix this by creating a new declaration group using an empty top-level splice:

data X = X
    deriving Eq

$(pure [])

$(do
    info <- reify ''X
    runIO $ print info
 )

We provide newDeclarationGroup as a means of documenting this behavior and providing a name for the pattern.

Since top level splices infer the presence of the $( ... ) brackets, we can also write:

data X = X
    deriving Eq

newDeclarationGroup

$(do
    info <- reify ''X
    runIO $ print info
 )

extsEnabled :: Q [Extension]

List all enabled language extensions.

isExtEnabled :: Extension -> Q Bool

Determine whether the given language extension is enabled in the Q monad.

lookupTypeName :: String -> Q (Maybe Name)

Look up the given name in the (type namespace of the) current splice's scope. See Language.Haskell.TH.Syntax for more details.

lookupValueName :: String -> Q (Maybe Name)

Look up the given name in the (value namespace of the) current splice's scope. See Language.Haskell.TH.Syntax for more details.

reifyFixity :: Name -> Q (Maybe Fixity)

reifyFixity nm attempts to find a fixity declaration for nm. For example, if the function foo has the fixity declaration infixr 7 foo, then reifyFixity 'foo would return Just (Fixity 7 InfixR). If the function bar does not have a fixity declaration, then reifyFixity 'bar returns Nothing, so you may assume bar has defaultFixity.

reifyType :: Name -> Q Type

reifyType nm attempts to find the type or kind of nm. For example, reifyType 'not returns Bool -> Bool, and reifyType ''Bool returns Type. This works even if there's no explicit signature and the type or kind is inferred.

reifyInstances :: Name -> [Type] -> Q [InstanceDec]

reifyInstances nm tys returns a list of all visible instances (see below for "visible") of nm tys. That is, if nm is the name of a type class, then all instances of this class at the types tys are returned. Alternatively, if nm is the name of a data family or type family, all instances of this family at the types tys are returned.

Note that this is a "shallow" test; the declarations returned merely have instance heads which unify with nm tys, they need not actually be satisfiable.

reifyInstances ''Eq [ TupleT 2 `AppT` ConT ''A `AppT` ConT ''B ] contains the instance (Eq a, Eq b) => Eq (a, b) regardless of whether A and B themselves implement Eq

reifyInstances ''Show [ VarT (mkName "a") ] produces every available instance of Eq

There is one edge case: reifyInstances ''Typeable tys currently always produces an empty list (no matter what tys are given).

In principle, the *visible* instances are * all instances defined in a prior top-level declaration group (see docs on newDeclarationGroup), or * all instances defined in any module transitively imported by the module being compiled

However, actually searching all modules transitively below the one being compiled is unreasonably expensive, so reifyInstances will report only the instance for modules that GHC has had some cause to visit during this compilation. This is a shortcoming: reifyInstances might fail to report instances for a type that is otherwise unusued, or instances defined in a different component. You can work around this shortcoming by explicitly importing the modules whose instances you want to be visible. GHC issue #20529 has some discussion around this.

isInstance :: Name -> [Type] -> Q Bool

Is the list of instances returned by reifyInstances nonempty?

If you're confused by an instance not being visible despite being defined in the same module and above the splice in question, see the docs for newDeclarationGroup for a possible explanation.

reifyRoles :: Name -> Q [Role]

reifyRoles nm returns the list of roles associated with the parameters (both visible and invisible) of the tycon nm. Fails if nm cannot be found or is not a tycon. The returned list should never contain InferR.

An invisible parameter to a tycon is often a kind parameter. For example, if we have

type Proxy :: forall k. k -> Type
data Proxy a = MkProxy

and reifyRoles Proxy, we will get [NominalR, PhantomR]. The NominalR is the role of the invisible k parameter. Kind parameters are always nominal.

reifyAnnotations :: Data a => AnnLookup -> Q [a]

reifyAnnotations target returns the list of annotations associated with target. Only the annotations that are appropriately typed is returned. So if you have Int and String annotations for the same target, you have to call this function twice.

reifyConStrictness :: Name -> Q [DecidedStrictness]

reifyConStrictness nm looks up the strictness information for the fields of the constructor with the name nm. Note that the strictness information that reifyConStrictness returns may not correspond to what is written in the source code. For example, in the following data declaration:

data Pair a = Pair a a

reifyConStrictness would return [DecidedLazy, DecidedLazy] under most circumstances, but it would return [DecidedStrict, DecidedStrict] if the -XStrictData language extension was enabled.

unTypeCode :: forall (r :: RuntimeRep) (a :: TYPE r) m. Quote m => Code m a -> m Exp

Extract the untyped representation from the typed representation

unsafeCodeCoerce :: forall (r :: RuntimeRep) (a :: TYPE r) m. Quote m => m Exp -> Code m a

Unsafely convert an untyped code representation into a typed code representation.

hoistCode :: forall m n (r :: RuntimeRep) (a :: TYPE r). Monad m => (forall x. m x -> n x) -> Code m a -> Code n a

Modify the ambient monad used during code generation. For example, you can use hoistCode to handle a state effect: handleState :: Code (StateT Int Q) a -> Code Q a handleState = hoistCode (flip runState 0)

bindCode :: forall m a (r :: RuntimeRep) (b :: TYPE r). Monad m => m a -> (a -> Code m b) -> Code m b

Variant of (>>=) which allows effectful computations to be injected into code generation.

bindCode_ :: forall m a (r :: RuntimeRep) (b :: TYPE r). Monad m => m a -> Code m b -> Code m b

Variant of (>>) which allows effectful computations to be injected into code generation.

joinCode :: forall m (r :: RuntimeRep) (a :: TYPE r). Monad m => m (Code m a) -> Code m a

A useful combinator for embedding monadic actions into Code myCode :: ... => Code m a myCode = joinCode $ do x <- someSideEffect return (makeCodeWith x)

liftCode :: forall (r :: RuntimeRep) (a :: TYPE r) m. m (TExp a) -> Code m a

Lift a monadic action producing code into the typed Code representation

mkName :: String -> Name

Generate a capturable name. Occurrences of such names will be resolved according to the Haskell scoping rules at the occurrence site.

For example:

f = [| pi + $(varE (mkName "pi")) |]
...
g = let pi = 3 in $f

In this case, g is desugared to

g = Prelude.pi + 3

Note that mkName may be used with qualified names:

mkName "Prelude.pi"

See also dyn for a useful combinator. The above example could be rewritten using dyn as

f = [| pi + $(dyn "pi") |]

nameBase :: Name -> String

The name without its module prefix.

Examples

Expand

>>> nameBase ''Data.Either.Either
"Either"
>>> nameBase (mkName "foo")
"foo"
>>> nameBase (mkName "Module.foo")
"foo"

nameModule :: Name -> Maybe String

Module prefix of a name, if it exists.

Examples

Expand

>>> nameModule ''Data.Either.Either
Just "Data.Either"
>>> nameModule (mkName "foo")
Nothing
>>> nameModule (mkName "Module.foo")
Just "Module"

namePackage :: Name -> Maybe String

A name's package, if it exists.

Examples

Expand

>>> namePackage ''Data.Either.Either
Just "base"
>>> namePackage (mkName "foo")
Nothing
>>> namePackage (mkName "Module.foo")
Nothing

nameSpace :: Name -> Maybe NameSpace

Returns whether a name represents an occurrence of a top-level variable (VarName), data constructor (DataName), type constructor, or type class (TcClsName). If we can't be sure, it returns Nothing.

Examples

Expand

>>> nameSpace 'Prelude.id
Just VarName
>>> nameSpace (mkName "id")
Nothing -- only works for top-level variable names
>>> nameSpace 'Data.Maybe.Just
Just DataName
>>> nameSpace ''Data.Maybe.Maybe
Just TcClsName
>>> nameSpace ''Data.Ord.Ord
Just TcClsName

tupleTypeName :: Int -> Name

Tuple type constructor

tupleDataName :: Int -> Name

Tuple data constructor

unboxedTupleTypeName :: Int -> Name

Unboxed tuple type constructor

unboxedTupleDataName :: Int -> Name

Unboxed tuple data constructor

unboxedSumTypeName :: SumArity -> Name

Unboxed sum type constructor

unboxedSumDataName :: SumAlt -> SumArity -> Name

Unboxed sum data constructor

defaultFixity :: Fixity

Default fixity: infixl 9

maxPrecedence :: Int

Highest allowed operator precedence for Fixity constructor (answer: 9)

putDoc :: DocLoc -> String -> Q ()

Add Haddock documentation to the specified location. This will overwrite any documentation at the location if it already exists. This will reify the specified name, so it must be in scope when you call it. If you want to add documentation to something that you are currently splicing, you can use addModFinalizer e.g.

do
  let nm = mkName "x"
  addModFinalizer $ putDoc (DeclDoc nm) "Hello"
  [d| $(varP nm) = 42 |]

The helper functions withDecDoc and withDecsDoc will do this for you, as will the funD_doc and other _doc combinators. You most likely want to have the -haddock flag turned on when using this. Adding documentation to anything outside of the current module will cause an error.

getDoc :: DocLoc -> Q (Maybe String)

Retreives the Haddock documentation at the specified location, if one exists. It can be used to read documentation on things defined outside of the current module, provided that those modules were compiled with the -haddock flag.

showName' :: NameIs -> Name -> String

dataToQa :: forall m a k q. (Quote m, Data a) => (Name -> k) -> (Lit -> m q) -> (k -> [m q] -> m q) -> (forall b. Data b => b -> Maybe (m q)) -> a -> m q

dataToQa is an internal utility function for constructing generic conversion functions from types with Data instances to various quasi-quoting representations. See the source of dataToExpQ and dataToPatQ for two example usages: mkCon, mkLit and appQ are overloadable to account for different syntax for expressions and patterns; antiQ allows you to override type-specific cases, a common usage is just const Nothing, which results in no overloading.

dataToExpQ :: (Quote m, Data a) => (forall b. Data b => b -> Maybe (m Exp)) -> a -> m Exp

dataToExpQ converts a value to a Exp representation of the same value, in the SYB style. It is generalized to take a function override type-specific cases; see liftData for a more commonly used variant.

dataToPatQ :: (Quote m, Data a) => (forall b. Data b => b -> Maybe (m Pat)) -> a -> m Pat

dataToPatQ converts a value to a Pat representation of the same value, in the SYB style. It takes a function to handle type-specific cases, alternatively, pass const Nothing to get default behavior.

newNameIO :: String -> IO Name

badIO :: String -> IO a

counter :: IORef Uniq

mkNameU :: String -> Uniq -> Name

Only used internally

unTypeQ :: forall (r :: RuntimeRep) (a :: TYPE r) m. Quote m => m (TExp a) -> m Exp

Discard the type annotation and produce a plain Template Haskell expression

Representation-polymorphic since template-haskell-2.16.0.0.

unsafeTExpCoerce :: forall (r :: RuntimeRep) (a :: TYPE r) m. Quote m => m Exp -> m (TExp a)

Annotate the Template Haskell expression with a type

This is unsafe because GHC cannot check for you that the expression really does have the type you claim it has.

Representation-polymorphic since template-haskell-2.16.0.0.

lookupName :: Bool -> String -> Q (Maybe Name)

getPackageRoot :: Q FilePath

Get the package root for the current package which is being compiled. This can be set explicitly with the -package-root flag but is normally just the current working directory.

The motivation for this flag is to provide a principled means to remove the assumption from splices that they will be executed in the directory where the cabal file resides. Projects such as haskell-language-server can't and don't change directory when compiling files but instead set the -package-root flag appropiately.

makeRelativeToProject :: FilePath -> Q FilePath

The input is a filepath, which if relative is offset by the package root.

addDependentFile :: FilePath -> Q ()

Record external files that runIO is using (dependent upon). The compiler can then recognize that it should re-compile the Haskell file when an external file changes.

Expects an absolute file path.

Notes:

ghc -M does not know about these dependencies - it does not execute TH.

The dependency is based on file content, not a modification time

addTempFile :: String -> Q FilePath

Obtain a temporary file path with the given suffix. The compiler will delete this file after compilation.

addTopDecls :: [Dec] -> Q ()

Add additional top-level declarations. The added declarations will be type checked along with the current declaration group.

addForeignFile :: ForeignSrcLang -> String -> Q ()

Deprecated: Use addForeignSource instead

addForeignSource :: ForeignSrcLang -> String -> Q ()

Emit a foreign file which will be compiled and linked to the object for the current module. Currently only languages that can be compiled with the C compiler are supported, and the flags passed as part of -optc will be also applied to the C compiler invocation that will compile them.

Note that for non-C languages (for example C++) extern C directives must be used to get symbols that we can access from Haskell.

To get better errors, it is recommended to use #line pragmas when emitting C files, e.g.

{-# LANGUAGE CPP #-}
...
addForeignSource LangC $ unlines
  [ "#line " ++ show (819 + 1) ++ " " ++ show "libraries/template-haskell/Language/Haskell/TH/Syntax.hs"
  , ...
  ]

addForeignFilePath :: ForeignSrcLang -> FilePath -> Q ()

Same as addForeignSource, but expects to receive a path pointing to the foreign file instead of a String of its contents. Consider using this in conjunction with addTempFile.

This is a good alternative to addForeignSource when you are trying to directly link in an object file.

addModFinalizer :: Q () -> Q ()

Add a finalizer that will run in the Q monad after the current module has been type checked. This only makes sense when run within a top-level splice.

The finalizer is given the local type environment at the splice point. Thus reify is able to find the local definitions when executed inside the finalizer.

addCorePlugin :: String -> Q ()

Adds a core plugin to the compilation pipeline.

addCorePlugin m has almost the same effect as passing -fplugin=m to ghc in the command line. The major difference is that the plugin module m must not belong to the current package. When TH executes, it is too late to tell the compiler that we needed to compile first a plugin module in the current package.

getQ :: Typeable a => Q (Maybe a)

Get state from the Q monad. Note that the state is local to the Haskell module in which the Template Haskell expression is executed.

putQ :: Typeable a => a -> Q ()

Replace the state in the Q monad. Note that the state is local to the Haskell module in which the Template Haskell expression is executed.

sequenceQ :: forall m. Monad m => forall a. [m a] -> m [a]

trueName :: Name

falseName :: Name

addrToByteArrayName :: Name

mkNameG_v :: String -> String -> String -> Name

addrToByteArray :: Int -> Addr# -> ByteArray

nothingName :: Name

justName :: Name

leftName :: Name

rightName :: Name

liftString :: Quote m => String -> m Exp

nonemptyName :: Name

mkNameG :: NameSpace -> String -> String -> String -> Name

Used for 'x etc, but not available to the programmer

oneName :: Name

manyName :: Name

mkOccName :: String -> OccName

mkPkgName :: String -> PkgName

mkModName :: String -> ModName

mkNameG_d :: String -> String -> String -> Name

showName :: Name -> String

liftData :: (Quote m, Data a) => a -> m Exp

liftData is a variant of lift in the Lift type class which works for any type with a Data instance.

modString :: ModName -> String

pkgString :: PkgName -> String

occString :: OccName -> String

thenCmp :: Ordering -> Ordering -> Ordering

mkNameL :: String -> Uniq -> Name

Only used internally

mkNameQ :: String -> String -> Name

Only used internally

mkNameS :: String -> Name

mkNameG_tc :: String -> String -> String -> Name

mk_tup_name :: Int -> NameSpace -> Bool -> Name

eqBytes :: Bytes -> Bytes -> Bool

compareBytes :: Bytes -> Bytes -> Ordering

memcmp :: Ptr a -> Ptr b -> CSize -> IO CInt

cmpEq :: Ordering -> Bool

Language extensions

module Language.Haskell.TH.LanguageExtensions

data ForeignSrcLang

Foreign formats supported by GHC via TH

Constructors

LangC
C

LangCxx
C++

LangObjc
Objective C

LangObjcxx
Objective C++

LangAsm
Assembly language (.s)

LangJs
JavaScript

RawObject
Object (.o)

Notes

Unresolved Infix

When implementing antiquotation for quasiquoters, one often wants to parse strings into expressions:

parse :: String -> Maybe Exp

But how should we parse a + b * c? If we don't know the fixities of + and *, we don't know whether to parse it as a + (b * c) or (a + b) * c.

In cases like this, use UInfixE, UInfixP, UInfixT, or PromotedUInfixT, which stand for "unresolved infix expressionpatterntype/promoted constructor", respectively. When the compiler is given a splice containing a tree of UInfixE applications such as

UInfixE
  (UInfixE e1 op1 e2)
  op2
  (UInfixE e3 op3 e4)

it will look up and the fixities of the relevant operators and reassociate the tree as necessary.

trees will not be reassociated across ParensE, ParensP, or ParensT, which are of use for parsing expressions like

(a + b * c) + d * e

InfixE, InfixP, InfixT, and PromotedInfixT expressions are never reassociated.

The UInfixE constructor doesn't support sections. Sections such as (a *) have no ambiguity, so InfixE suffices. For longer sections such as (a + b * c -), use an InfixE constructor for the outer-most section, and use UInfixE constructors for all other operators:

InfixE
  Just (UInfixE ...a + b * c...)
  op
  Nothing

Sections such as (a + b +) and ((a + b) +) should be rendered into Exps differently:

(+ a + b)   ---> InfixE Nothing + (Just $ UInfixE a + b)
                   -- will result in a fixity error if (+) is left-infix
(+ (a + b)) ---> InfixE Nothing + (Just $ ParensE $ UInfixE a + b)
                   -- no fixity errors

Quoted expressions such as

[| a * b + c |] :: Q Exp
[p| a : b : c |] :: Q Pat
[t| T + T |] :: Q Type

will never contain UInfixE, UInfixP, UInfixT, PromotedUInfixT, InfixT, 'PromotedInfixT, ParensE, ParensP, or ParensT constructors.